--- portal-web/test/com/liferay/portalweb/plugins/todaysevents/portlet/addportletduplicate/AddPortletTETest.html
@@ -1,59 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>AddPortletTETest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">AddPortletTETest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/web/guest/home/</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>link=Today's Events Test Page</td>
-	<td>Today's Events Test Page</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Today's Events Test Page</td>
-	<td>Today's Events Test Page</td>
-</tr>
-<tr>
-	<td>assertPartialText</td>
-	<td>//a[@id='_145_addApplication']</td>
-	<td>More</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//a[@id='_145_addApplication']</td>
-	<td>More</td>
-</tr>
-<tr>
-	<td>waitForElementPresent</td>
-	<td>//div[@title="Today's Events"]/p/a</td>
-	<td>Add</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//div[@title="Today's Events"]/p/a</td>
-	<td>Add</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//section</td>
-	<td>Portlet</td>
-</tr>
-<tr>
-	<td>assertVisible</td>
-	<td>//section</td>
-	<td>Portlet</td>
-</tr>
-
-</tbody></table>
-</body>
-</html>
