--- portal-web/test/com/liferay/portalweb/plugins/todaysevents/portlet/removeportlet/AddPageTETest.html
@@ -1,59 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>AddPageTETest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">AddPageTETest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/web/guest/home/</td>
-	<td></td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//nav[@id='navigation']</td>
-	<td>Navigation</td>
-</tr>
-<tr>
-	<td>waitForElementPresent</td>
-	<td>//a[@id='addPage']</td>
-	<td>Add Page</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//a[@id='addPage']</td>
-	<td>Add Page</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//input[@type='text']</td>
-	<td>Page Name</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@type='text']</td>
-	<td>Today's Events Test Page</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//button[@id='save']</td>
-	<td>Save</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>link=Today's Events Test Page</td>
-	<td>Today's Events Test Page</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Today's Events Test Page</td>
-	<td>Today's Events Test Page</td>
-</tr>
-
-</tbody></table>
-</body>
-</html>
