--- portal-web/test/com/liferay/portalweb/plugins/todaysevents/portlet/removeportlet/RemovePortletTETest.html
@@ -1,42 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>RemovePortletTETest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">RemovePortletTETest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/web/guest/home/</td>
-	<td></td>
-</tr><tr>
-	<td>waitForVisible</td>
-	<td>link=Today's Events Test Page</td>
-	<td>Today's Events Test Page</td>
-</tr><tr>
-	<td>clickAtAndWait</td>
-	<td>link=Today's Events Test Page</td>
-	<td>Today's Events Test Page</td>
-</tr><tr>
-	<td>click</td>
-	<td>//img[@alt='Remove']</td>
-	<td>Remove</td>
-</tr><tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to remove this component?</td>
-	<td></td>
-</tr><tr>
-	<td>waitForElementNotPresent</td>
-	<td>//section</td>
-	<td>Portlet</td>
-</tr><tr>
-	<td>assertElementNotPresent</td>
-	<td>//section</td>
-	<td>Portlet</td>
-</tr>
-</tbody></table>
-</body>
-</html>
