--- portal-web/test/com/liferay/portalweb/plugins/todaysevents/portlet/removeportlet/RemovePortletTests.html
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-  <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>RemovePortletTests</title>
-</head>
-<body>
-<table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>RemovePortletTests</b></td></tr>
-<tr><td><a href="AddPageTETest.html">AddPageTETest</a></td></tr>
-<tr><td><a href="AddPortletTETest.html">AddPortletTETest</a></td></tr>
-<tr><td><a href="RemovePortletTETest.html">RemovePortletTETest</a></td></tr>
-<tr><td><a href="TearDownPageTest.html">TearDownPageTest</a></td></tr>
-</tbody></table>
-</body>
-</html>
